improvement switch to BigDecimal, hide unnecessary zeros in price etc.

+++ src/main/java/com/cryptotrading/Crypto/Trading/App/service/UserService.java
@@ -4,17 +4,18 @@ import com.cryptotrading.Crypto.Trading.App.model.dto.UserRegisterBindingModel;
 import com.cryptotrading.Crypto.Trading.App.model.entity.Transaction;
 import com.cryptotrading.Crypto.Trading.App.model.entity.User;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface UserService {
     boolean register(UserRegisterBindingModel userRegisterBindingModel);
-    Optional<Transaction> buyCrypto(String username, String pair, double spend, double quantity);
+    Optional<Transaction> buyCrypto(String username, String pair, BigDecimal spend, BigDecimal quantity);
 
-    double getBalance(String email);
+    BigDecimal getBalance(String email);
 
-    double getQuantityFromPair(String email, String pair);
+    BigDecimal getQuantityFromPair(String email, String pair);
 
-    Optional<Transaction> sellCrypto(String email, String pair, double spend, double quantity);
+    Optional<Transaction> sellCrypto(String email, String pair, BigDecimal spend, BigDecimal quantity);
     User findByEmail(String email);
 
     void resetProfile(String email);
